requirements.txt
flask==3.0.0
litellm==1.16.8
python-dotenv==1.0.0